backup_automation/requirements.txt
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
pyzipper==0.3.6
schedule==1.2.1
PyYAML==6.0.1
